refactor: dao 내 예외 처리 메시지 추가
